hi送信ロジック
